requirements.txt
# Base Dependencies
pandas
requests

# Google Cloud SDKs 
google-cloud-storage
google-cloud-bigquery

# # dbt Core and BigQuery Adapter
dbt-core==1.8.8
dbt-bigquery==1.8.0

# Airflow provider for GCP
apache-airflow-providers-google